Move save stat functionality to overview
